feat: changes database structure to allow games by account logged in

+++ src/services/database.ts
@@ -1,8 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable class-methods-use-this */
+import useAuth from '../hooks/useAuth';
 import { firestore } from './firebase';
 
-const db = firestore.collection('/games');
+const db = firestore.collection('games');
 
 class FirestoreService {
   async getAll() {
@@ -10,36 +11,47 @@ class FirestoreService {
     return snapshot.docs.map(document => document.data());
   }
 
-  async getOne(id: number | undefined) {
-    const snapshotID: number[] = [];
+  async getOne(id: number | undefined, userID: any) {
+    const snapshotID: any = [];
 
     const snapshot = await db
       .where('id', '==', id)
       .get()
       .then(querySnapshot => {
         querySnapshot.forEach(doc => {
-          return snapshotID.push(doc.data().id);
+          if (doc.data().userId === userID) {
+            snapshotID.push(doc.data().id);
+          }
         });
       });
-
     return snapshotID;
   }
 
-  create(value) {
-    return db.add(value);
+  async create({ id, name }, userId: any) {
+    const user = await firestore
+      .collection('user_info')
+      .doc(userId)
+      .set({ id: userId });
+
+    const games = await db.add({ id, name, userId });
+
+    const data = { games, user };
+    return data;
   }
 
   update(id, value) {
     return db.doc(id).update(value);
   }
 
-  async delete(id: number) {
+  async delete(id: number, userID) {
     const snapshot = await db
       .where('id', '==', id)
       .get()
       .then(querySnapshot => {
         querySnapshot.forEach(doc => {
-          return db.doc(doc.id).delete();
+          if (doc.data().userId === userID) {
+            db.doc(doc.id).delete();
+          }
         });
       });
